Refactor: improved clinical entity controller and service with proper authentication

+++ personal-health-passport/Controllers/ClinicalEntityController.cs
@@ -3,11 +3,15 @@ using personal_health_passport.DTOs;
 using personal_health_passport.Models;
 using personal_health_passport.Services;
 using System.Net.Http;
+using System.Security.Claims;
+using Microsoft.AspNetCore.Authorization;
+
 
 namespace personal_health_passport.Controllers
 {
     [Route("nlp")]
     [ApiController]
+    [Authorize]
     public class ClinicalEntityController : Controller
     {
         private readonly HttpClient Http;
@@ -17,6 +21,16 @@ namespace personal_health_passport.Controllers
         {
             _entityService = entityService;
             Http = http;
+        }
+
+        private string? GetLoggedInUserId()
+        {
+            //Get the UserId from the token, if automatic translation is off it will fallback to using "sub" to find UserId
+            //returns null if sub does not exisit 
+            var claim = User.FindFirst(ClaimTypes.NameIdentifier)?.Value
+                      ?? User.FindFirst("sub")?.Value;
+
+            return claim;
         }
 
         [HttpPost("generate")]
@@ -48,7 +62,13 @@ namespace personal_health_passport.Controllers
                     return BadRequest("The NLP service returned an empty response.");
                 }
 
-                if (_entityService.AddEntitiesToDb(result))
+                string userId = GetLoggedInUserId();
+                if(userId == null)
+                {
+                    return Unauthorized();
+                }
+
+                if (_entityService.AddEntitiesToDb(userId,result))
                 {
                     return Ok(result);
                 }
@@ -68,6 +88,7 @@ namespace personal_health_passport.Controllers
         }
 
         [HttpGet("{id}")]
+        [Authorize(Roles = "Admin")]
         public IActionResult GetEntity(int id)
         {
             var entity = _entityService.GetEntity(id);
@@ -79,8 +100,10 @@ namespace personal_health_passport.Controllers
         }
 
         [HttpGet("user")]
-        public IActionResult GetAllEntitiesByUser(string? userId)
+        public IActionResult GetAllEntitiesByUser()
         {
+
+            string? userId = GetLoggedInUserId();
             var entities = _entityService.GetAllEntitiesByUser(userId);
 
             return Ok(entities);
@@ -89,16 +112,37 @@ namespace personal_health_passport.Controllers
         [HttpPost]
         public IActionResult AddEntity([FromBody] ClinicalEntity entity)
         {
+            string? userId = GetLoggedInUserId();
+            
+            if(userId != null)
+            {
+                entity.Uid = userId;
+            }
+            else
+            {
+                return Unauthorized();
+            }
+
             var result = _entityService.AddEntity(entity);
 
             return Ok(result);
         }
 
         [HttpPut("{id}")]
-        public IActionResult UpdateEntity(
-            int id,
-            [FromBody] ClinicalEntity updatedEntity)
+        public IActionResult UpdateEntity(int id,[FromBody] ClinicalEntity updatedEntity)
         {
+            string? userId = GetLoggedInUserId();
+
+            if (userId == null || updatedEntity.Uid != userId )
+            {
+                return Unauthorized();
+            }
+
+            if(id != updatedEntity.Id)
+            {
+                return BadRequest();
+            }
+
             var result = _entityService.UpdateEntity(id, updatedEntity);
 
             if (result == null)
@@ -110,7 +154,24 @@ namespace personal_health_passport.Controllers
         [HttpDelete("{id}")]
         public IActionResult DeleteEntity(int id)
         {
+
+            string? userId = GetLoggedInUserId();
+            ClinicalEntity entity = _entityService.GetEntity(id);
+
+
+            if (entity == null)
+            {
+                return BadRequest();
+            }
+
+            if (userId == null || entity.Uid != userId)
+            {
+                return Unauthorized();
+            }
+
+
             var deleted = _entityService.DeleteEntity(id);
+
 
             if (!deleted)
                 return NotFound();
@@ -119,8 +180,7 @@ namespace personal_health_passport.Controllers
         }
 
         [HttpDelete]
-        public IActionResult DeleteEntities(
-            [FromBody] List<ClinicalEntity> entities)
+        public IActionResult DeleteEntities([FromBody] List<ClinicalEntity> entities)
         {
             _entityService.DeleteEntities(entities);
